requirements.txt
matplotlib
multiprocess
numpy
pandas
pickleshare
ruamel.yaml
scikit-learn
scipy
seaborn
tqdm
